refactor(Tabs): Refactor Tabs so the Tab Button is rendered in a seperate component
#3032
