+++ packages/patternfly-4/react-core/src/components/Tabs/Tab.test.tsx
@@ -2,16 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { Tab } from './Tab';
 
-test('should render tab', () => {
-  const view = shallow(
-    <Tab eventKey={0} title="Tab item 1">
-      Tab 1 section
-    </Tab>
-  );
-  expect(view).toMatchSnapshot();
-});
-
-test('should render active tab', () => {
+test('should not render anything', () => {
   const view = shallow(
     <Tab eventKey={0} title="Tab item 1">
       Tab 1 section
