+++ packages/patternfly-4/react-core/src/components/Tabs/TabButton.test.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { TabButton } from './TabButton';
+
+test('should render tab button', () => {
+  const view = shallow(
+    <TabButton eventKey={0} >
+      Tab button
+    </TabButton>
+  );
+  expect(view).toMatchSnapshot();
+});
