+++ packages/patternfly-4/react-core/src/components/Tabs/Tabs.tsx
@@ -7,6 +7,7 @@ import { AngleLeftIcon, AngleRightIcon } from '@patternfly/react-icons';
 import { getUniqueId, isElementInView, sideElementIsOutOfView } from '../../helpers/util';
 import { SIDE } from '../../helpers/constants';
 import { Tab } from './Tab';
+import { TabButton } from './TabButton';
 import { TabContent } from './TabContent';
 import { InjectedOuiaProps, withOuiaContext } from '../withOuia';
 
@@ -244,27 +245,24 @@ class Tabs extends React.Component<TabsProps & InjectedOuiaProps, TabsState> {
           </button>
           <ul className={css(styles.tabsList)} ref={this.tabList} onScroll={this.handleScrollButtons}>
             {React.Children.map(children, (child: any, index) => {
-              const { title, eventKey, tabContentRef, id: childId, tabContentId, ...rest } = child.props;
+              const { title, eventKey, tabContentRef, id: childId, tabContentId, className, ...rest } = child.props;
               return (
                 <li
                   key={index}
                   className={css(styles.tabsItem, eventKey === activeKey && styles.modifiers.current, className)}
                 >
-                  <Tab
-                    className={css(styles.tabsButton)}
+                  <TabButton
+                    className={css(styles.tabsButton, className)}
                     onClick={(event: any) => this.handleTabClick(event, eventKey, tabContentRef, mountOnEnter)}
                     id={`pf-tab-${eventKey}-${childId || uniqueId}`}
                     aria-controls={
                       tabContentId ? `${tabContentId}` : `pf-tab-section-${eventKey}-${childId || uniqueId}`
                     }
-                    tabContentId={tabContentId}
                     tabContentRef={tabContentRef}
-                    eventKey={eventKey}
-                    title={title}
                     {...rest}
                   >
                     {title}
-                  </Tab>
+                  </TabButton>
                 </li>
               );
             })}
